+++ xfel/merging/application/prepare/prepare_spread.py
@@ -5,6 +5,109 @@ from xfel.merging.application.worker import worker
 from dials.array_family import flex
 
 
+# Template for individual slice phil files
+SLICE_PHIL_TEMPLATE = """\
+# Base phil from: {stage2_phil_path}
+{base_phil_content}
+
+# Slice {slice_idx:03d}: percentiles {pct_start}-{pct_end}, center {center}
+# Energy range: {energy_start:.2f} - {energy_end:.2f} eV
+
+{input_paths}
+input.experiments_suffix=.expt
+input.reflections_suffix=.refl
+output.output_dir={slice_output_dir}
+"""
+
+# Template for the unified batch script (SLURM + local)
+SPREAD_SCRIPT_TEMPLATE = """\
+#!/bin/bash
+#SBATCH --array=0-{n_slices}
+#SBATCH --time={slurm_time_limit}
+{slurm_optional_directives}#SBATCH --job-name=spread
+#SBATCH --output={scripts_dir}/slurm_%A_%a.out
+#SBATCH --error={scripts_dir}/slurm_%A_%a.err
+
+# This script works in two modes:
+# 1. SLURM: sbatch run_spread.sh (each array task runs one iteration)
+# 2. Local: bash run_spread.sh (runs all tasks sequentially)
+
+N_SLICES={n_slices}
+SCRIPTS_DIR="{scripts_dir}"
+STATUS_DIR="{status_dir}"
+STAGE2_DIR="{stage2_output_dir}"
+MTZ_NAME="{mtz_name}"
+NPROC={stage2_nproc}
+PDB_FILE="{phenix_pdb_path}"
+PHENIX_PHIL="{phenix_phil_path}"
+CCTBX_ACTIVATE="{cctbx_activate}"
+PHENIX_ACTIVATE="{phenix_activate}"
+
+mkdir -p ${{STATUS_DIR}}
+
+# Determine task IDs based on execution mode
+if [ -n $SLURM_ARRAY_TASK_ID ]; then
+  TASK_IDS=($SLURM_ARRAY_TASK_ID)
+  RUN_CMD="srun -n ${{NPROC}}"
+else
+  TASK_IDS=($(seq 0 ${{N_SLICES}}))
+  RUN_CMD="mpirun -n ${{NPROC}}"
+fi
+
+for TASK_ID in ${{TASK_IDS[@]}}; do
+  if [ $TASK_ID -lt $N_SLICES ]; then
+    # Stage 2 merge task
+    source $CCTBX_ACTIVATE
+    SLICE_IDX=$(printf "%03d" $TASK_ID)
+    PHIL_FILE="${{SCRIPTS_DIR}}/stage2_slice_${{SLICE_IDX}}.phil"
+    echo "Running stage 2 merge for slice ${{SLICE_IDX}}..."
+    ${{RUN_CMD}} cctbx.xfel.merge ${{PHIL_FILE}}
+
+    # Write status file on completion
+    touch ${{STATUS_DIR}}/slice_${{SLICE_IDX}}.done
+  else
+    # Phenix refinement coordinator task
+
+    # In SLURM mode, poll for all merges to complete
+    if [ -n $SLURM_ARRAY_TASK_ID ]; then
+      echo "Waiting for all stage 2 merge tasks to complete..."
+      while true; do
+        DONE_COUNT=$(ls ${{STATUS_DIR}}/slice_*.done 2>/dev/null | wc -l)
+        if [ $DONE_COUNT -ge $N_SLICES ]; then
+          echo "All $N_SLICES merge tasks complete."
+          break
+        fi
+        echo "Waiting... $DONE_COUNT / $N_SLICES complete"
+        sleep 5
+      done
+    fi
+
+    # Activate phenix environment and run all refinements in parallel
+    source $PHENIX_ACTIVATE
+    echo "Starting phenix refinements..."
+    for i in $(seq 0 $((N_SLICES - 1))); do
+      SLICE_IDX=$(printf "%03d" $i)
+      SLICE_DIR="${{STAGE2_DIR}}/slice_${{SLICE_IDX}}"
+      MTZ_FILE="${{SLICE_DIR}}/${{MTZ_NAME}}"
+
+      if [ -f "${{MTZ_FILE}}" ]; then
+        (
+          cd "${{SLICE_DIR}}"
+          phenix.refine ${{MTZ_FILE}} ${{PDB_FILE}} ${{PHENIX_PHIL}} output.prefix=refine_${{SLICE_IDX}} > refine_${{SLICE_IDX}}.log 2>&1
+        ) &
+      else
+        echo "Warning: ${{MTZ_FILE}} not found, skipping slice ${{SLICE_IDX}}"
+      fi
+    done
+
+    echo "Waiting for all phenix refinements to complete..."
+    wait
+    echo "All phenix refinements complete."
+  fi
+done
+"""
+
+
 class prepare_spread(worker):
   """
   Prepare data for SpReAD (Spectral Resolved Anomalous Diffraction) analysis.
@@ -16,7 +119,7 @@ class prepare_spread(worker):
   """
 
   def __init__(self, params, mpi_helper=None, mpi_logger=None):
-    super(spread, self).__init__(params=params, mpi_helper=mpi_helper, mpi_logger=mpi_logger)
+    super(prepare_spread, self).__init__(params=params, mpi_helper=mpi_helper, mpi_logger=mpi_logger)
 
   def __repr__(self):
     return "Prepare SpReAD energy-binned datasets"
@@ -24,10 +127,9 @@ class prepare_spread(worker):
   def run(self, experiments, reflections):
     self.logger.log_step_time("SPREAD_PREPARE")
 
-    # Extract parameters (with defaults if not yet in phil)
-    n_bins = getattr(self.params.prepare.spread, 'n_energy_bins', 100)
-    output_dir = getattr(self.params.prepare.spread, 'output_dir',
-                         self.params.output.output_dir)
+    # Extract parameters
+    n_bins = self.params.prepare.spread.n_energy_bins
+    output_dir = self.params.prepare.spread.output_dir or self.params.output.output_dir
 
     # Step 1: Compute local energies from wavelengths
     self.logger.log_step_time("COMPUTE_ENERGIES")
@@ -62,10 +164,16 @@ class prepare_spread(worker):
                              expt_to_bin, n_bins, output_dir)
     self.logger.log_step_time("WRITE_FILES", True)
 
+    # Step 6: Write batch scripts (only on rank 0)
+    if self.mpi_helper.rank == 0:
+      self.logger.log_step_time("WRITE_SCRIPTS")
+      self._write_batch_scripts(n_bins, bin_edges, output_dir)
+      self.logger.log_step_time("WRITE_SCRIPTS", True)
+
     self.logger.log_step_time("SPREAD_PREPARE", True)
 
-    # Return empty - this worker is a terminal step that writes to disk
-    return None, None
+    # Return original experiments and reflections for potential downstream processing
+    return experiments, reflections
 
   def _compute_local_energies(self, experiments):
     """
@@ -268,3 +376,133 @@ class prepare_spread(worker):
       refls.as_file(refl_path)
 
       self.logger.log(f"Wrote {len(expts)} experiments to {filename_base}.expt/.refl")
+
+  def _write_batch_scripts(self, n_bins, bin_edges, output_dir):
+    """
+    Generate batch scripts for running stage 2 merging jobs.
+
+    Writes:
+      - stage2_slice_XXX.phil for each energy slice
+      - run_spread.sh for SLURM/local execution
+      - slices.txt metadata file
+    """
+    # Extract parameters
+    window_width = self.params.prepare.spread.window_width
+    window_step = self.params.prepare.spread.window_step
+    stage2_phil_path = self.params.prepare.spread.stage2_phil
+    stage2_nproc = self.params.prepare.spread.stage2_nproc
+    stage2_output_dir = self.params.prepare.spread.stage2_output_dir or os.path.join(output_dir, 'stage2')
+
+    # Phenix refinement parameters
+    phenix_phil_path = self.params.prepare.spread.phenix_phil or ""
+    phenix_pdb_path = self.params.prepare.spread.phenix_pdb or ""
+    mtz_name = self.params.prepare.spread.mtz_name
+
+    # SLURM parameters (optional)
+    slurm_partition = self.params.prepare.spread.slurm_partition
+    slurm_account = self.params.prepare.spread.slurm_account
+    slurm_time_limit = self.params.prepare.spread.slurm_time_limit
+    slurm_constraint = self.params.prepare.spread.slurm_constraint
+
+    # Activation scripts
+    cctbx_activate = self.params.prepare.spread.cctbx_activate
+    phenix_activate = self.params.prepare.spread.phenix_activate
+
+    # Read base phil content if provided
+    base_phil_content = ""
+    if stage2_phil_path and os.path.exists(stage2_phil_path):
+      with open(stage2_phil_path, 'r') as f:
+        base_phil_content = f.read()
+
+    # Calculate valid slice centers
+    # Window covers [center - width/2, center + width/2)
+    # For 20% window: first valid center is 10 (covers 0-19), last is 90 (covers 80-99)
+    half_width = window_width // 2
+    first_center = half_width
+    last_center = (n_bins - 1) - half_width + (1 if window_width % 2 == 0 else 0)
+
+    slice_centers = list(range(first_center, last_center + 1, window_step))
+    n_slices = len(slice_centers)
+
+    self.logger.log(f"Generating {n_slices} slice phil files (window_width={window_width}%, step={window_step}%)")
+
+    # Create scripts directory
+    scripts_dir = os.path.join(output_dir, 'scripts')
+    status_dir = os.path.join(scripts_dir, 'status')
+    os.makedirs(scripts_dir, exist_ok=True)
+    os.makedirs(stage2_output_dir, exist_ok=True)
+
+    # Write slices.txt metadata
+    slices_metadata_path = os.path.join(scripts_dir, 'slices.txt')
+    with open(slices_metadata_path, 'w') as f:
+      f.write("# slice_index center_percentile pct_start pct_end energy_start_eV energy_end_eV\n")
+      for slice_idx, center in enumerate(slice_centers):
+        pct_start = center - half_width
+        pct_end = pct_start + window_width - 1
+        energy_start = bin_edges[pct_start]
+        energy_end = bin_edges[pct_end + 1]
+        f.write(f"{slice_idx:03d} {center:3d} {pct_start:3d} {pct_end:3d} {energy_start:.2f} {energy_end:.2f}\n")
+
+    # Determine bin index width for filename formatting
+    bin_width = len(str(n_bins - 1))
+
+    # Write individual slice phil files
+    for slice_idx, center in enumerate(slice_centers):
+      pct_start = center - half_width
+      pct_end = pct_start + window_width - 1
+
+      # Build input paths
+      input_paths = "\n".join(
+        f"input.path={output_dir}/pct_{pct:0{bin_width}d}_c*.*"
+        for pct in range(pct_start, pct_end + 1)
+      )
+
+      slice_phil_content = SLICE_PHIL_TEMPLATE.format(
+        stage2_phil_path=stage2_phil_path or "N/A",
+        base_phil_content=base_phil_content,
+        slice_idx=slice_idx,
+        pct_start=pct_start,
+        pct_end=pct_end,
+        center=center,
+        energy_start=bin_edges[pct_start],
+        energy_end=bin_edges[pct_end + 1],
+        input_paths=input_paths,
+        slice_output_dir=os.path.join(stage2_output_dir, f'slice_{slice_idx:03d}')
+      )
+
+      slice_phil_path = os.path.join(scripts_dir, f'stage2_slice_{slice_idx:03d}.phil')
+      with open(slice_phil_path, 'w') as f:
+        f.write(slice_phil_content)
+
+    # Build optional SLURM directives
+    slurm_optional = ""
+    if slurm_partition:
+      slurm_optional += f"#SBATCH --partition={slurm_partition}\n"
+    if slurm_account:
+      slurm_optional += f"#SBATCH --account={slurm_account}\n"
+    if slurm_constraint:
+      slurm_optional += f"#SBATCH --constraint={slurm_constraint}\n"
+
+    # Write unified batch script
+    spread_script_content = SPREAD_SCRIPT_TEMPLATE.format(
+      n_slices=n_slices,
+      slurm_time_limit=slurm_time_limit,
+      slurm_optional_directives=slurm_optional,
+      scripts_dir=scripts_dir,
+      status_dir=status_dir,
+      stage2_output_dir=stage2_output_dir,
+      mtz_name=mtz_name,
+      stage2_nproc=stage2_nproc,
+      phenix_pdb_path=phenix_pdb_path,
+      phenix_phil_path=phenix_phil_path,
+      cctbx_activate=cctbx_activate,
+      phenix_activate=phenix_activate
+    )
+
+    spread_script_path = os.path.join(scripts_dir, 'run_spread.sh')
+    with open(spread_script_path, 'w') as f:
+      f.write(spread_script_content)
+    os.chmod(spread_script_path, 0o755)
+
+    self.logger.log(f"Wrote {n_slices} slice phil files to {scripts_dir}")
+    self.logger.log(f"Wrote batch script: run_spread.sh (use 'sbatch' for SLURM, 'bash' for local)")
